fix: 약관 로직 수정 - [선호항목 입력]화면에서 [학교인증]화면으로 순서 이동
